chore: move some types around
