+++ src/oauth/base.ts
@@ -1,0 +1,9 @@
+export interface IAuthorizationResponse {
+  readonly accessToken: string
+  readonly refreshToken?: string
+}
+
+export type TokenRequester = (
+  url: string,
+  params: IndexSignature,
+) => Promise<IAuthorizationResponse>
