+++ src/oauth/index.ts
@@ -3,24 +3,11 @@ import * as implicitGrant from './implicitGrant'
 import * as passwordGrant from './passwordGrant'
 import * as refreshTokenGrant from './refreshTokenGrant'
 
+export * from './base'
+
 export {
   authorizationCodeGrant,
   implicitGrant,
   passwordGrant,
   refreshTokenGrant,
 }
-
-export interface IAuthorizationResponse {
-  readonly accessToken: string
-  readonly refreshToken?: string
-}
-
-export type RequestTokenParams =
-  | authorizationCodeGrant.IAccessTokenRequestParams
-  | passwordGrant.IAccessTokenRequestParams
-  | refreshTokenGrant.IAccessTokenRequestParams
-
-export type TokenRequester = (
-  url: string,
-  params: IndexSignature,
-) => Promise<IAuthorizationResponse>
