+++ src/oauth/passwordGrant.ts
@@ -1,7 +1,7 @@
 import memoize from 'mem'
 
-import { TokenRequester } from '.'
 import { DEFAULT_MEMOIZE_OPTIONS } from '../constants'
+import { TokenRequester } from './base'
 
 export const GRANT_TYPE = 'password'
 
@@ -14,7 +14,7 @@ export interface IAccessTokenRequestParams {
   readonly scope?: string
 }
 
-const castToRequestParams = (
+const castToTokenRequestParams = (
   params: IndexSignature,
 ): IAccessTokenRequestParams => {
   const { username, password, scope, clientId, clientSecret } = params
@@ -49,7 +49,7 @@ const castToRequestParams = (
 
 export const isEligible = (params: IndexSignature): boolean => {
   try {
-    return !!castToRequestParams(params)
+    return !!castToTokenRequestParams(params)
   } catch {
     return false
   }
@@ -61,12 +61,12 @@ export const requestToken = memoize(
 
     return oauthTokenRequest(
       `${oauthUrl}/oauth/token`,
-      castToRequestParams(params),
+      castToTokenRequestParams(params),
     )
   },
   {
     ...DEFAULT_MEMOIZE_OPTIONS,
     cacheKey: (_: TokenRequester, params: IndexSignature) =>
-      JSON.stringify(castToRequestParams(params)),
+      JSON.stringify(castToTokenRequestParams(params)),
   },
 )
