+++ src/oauth/refreshTokenGrant.ts
@@ -1,7 +1,7 @@
 import memoize from 'mem'
 
-import { TokenRequester } from '.'
 import { DEFAULT_MEMOIZE_OPTIONS } from '../constants'
+import { TokenRequester } from './base'
 
 export const GRANT_TYPE = 'refresh_token'
 
@@ -13,7 +13,7 @@ export interface IAccessTokenRequestParams {
   readonly scope?: string
 }
 
-const castToRequestParams = (
+const castToTokenRequestParams = (
   params: IndexSignature,
 ): IAccessTokenRequestParams => {
   const { clientId, clientSecret, refreshToken, scope } = params
@@ -41,7 +41,7 @@ const castToRequestParams = (
 
 export const isEligible = (params: IndexSignature): boolean => {
   try {
-    return !!castToRequestParams(params)
+    return !!castToTokenRequestParams(params)
   } catch {
     return false
   }
@@ -53,12 +53,12 @@ export const requestToken = memoize(
 
     return oauthTokenRequest(
       `${oauthUrl}/oauth/token`,
-      castToRequestParams(params),
+      castToTokenRequestParams(params),
     )
   },
   {
     ...DEFAULT_MEMOIZE_OPTIONS,
     cacheKey: (_: TokenRequester, params: IndexSignature) =>
-      JSON.stringify(castToRequestParams(params)),
+      JSON.stringify(castToTokenRequestParams(params)),
   },
 )
